refactor(backend): standardize Post constructs names to Model, and Type
