+++ packages/backend/src/post/mutations/likePostMutation.ts
@@ -21,8 +21,8 @@ import {mutationWithClientMutationId} from 'graphql-relay';
 
 import TicoTecoContext from '../../graphql/TicoTecoContext';
 
-import Post from '../Post';
-import GraphQLPost from '../types/GraphQLPost';
+import Post from '../PostModel';
+import PostType from '../PostType';
 
 type LikePostArgs = {
   readonly id: string;
@@ -35,7 +35,7 @@ export const likePostMutation = mutationWithClientMutationId({
   },
   outputFields: () => ({
     post: {
-      type: new GraphQLNonNull(GraphQLPost),
+      type: new GraphQLNonNull(PostType),
       resolve: ({post}) => post,
     },
   }),
