+++ packages/backend/src/post/mutations/updatePostMutation.ts
@@ -23,8 +23,8 @@ import * as Yup from 'yup';
 
 import TicoTecoContext from '../../graphql/TicoTecoContext';
 
-import Post from '../Post';
-import GraphQLPost from '../types/GraphQLPost';
+import Post from '../PostModel';
+import PostType from '../PostType';
 
 type UpdatePostArgs = {
   readonly id: string;
@@ -46,7 +46,7 @@ export const updatePostMutation = mutationWithClientMutationId({
   },
   outputFields: () => ({
     post: {
-      type: new GraphQLNonNull(GraphQLPost),
+      type: new GraphQLNonNull(PostType),
       resolve: ({post}) => post,
     },
   }),
