+++ packages/backend/src/post/queries/postQuery.ts
@@ -20,8 +20,8 @@ import {GraphQLFieldConfig, GraphQLID} from 'graphql';
 
 import TicoTecoContext from '../../graphql/TicoTecoContext';
 
-import GraphQLPost from '../types/GraphQLPost';
-import Post from '../Post';
+import PostType from '../PostType';
+import Post from '../PostModel';
 
 export type PostArguments = {
   readonly id: string;
@@ -30,7 +30,7 @@ export type PostArguments = {
 type PostQuery = GraphQLFieldConfig<any, TicoTecoContext, PostArguments>;
 
 export const postQuery: PostQuery = {
-  type: GraphQLPost,
+  type: PostType,
   description: 'Finds a post by id',
   args: {
     id: {type: GraphQLID},
